added new files with new screens and change the main screen
