+++ Snowboard/app/src/main/res/drawable/gradient_background.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item>
+        <shape>
+            <gradient
+                android:angle="315"
+                android:endColor="#6667AB"
+                android:startColor="#F5D5E0"
+                android:type="linear" />
+        </shape>
+    </item>
+
+</selector>
