+++ Snowboard/app/src/main/res/drawable/ic_equipment.xml
@@ -1,0 +1,75 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="M217.04,506.43H16.7C7.47,506.43 0,498.96 0,489.74V389.57c0,-46.1 37.38,-83.48 83.48,-83.48h244.87c9.22,0 16.7,7.47 16.7,16.7v66.78c0,9.22 -7.47,16.7 -16.7,16.7h-85.43c-5.07,0 -9.18,4.11 -9.18,9.18v74.3C233.74,498.96 226.26,506.43 217.04,506.43z"
+      android:fillColor="#99DDFF"/>
+  <path
+      android:pathData="M328.35,306.09H150.26v200.35h66.78c9.22,0 16.7,-7.47 16.7,-16.7v-74.3c0,-5.07 4.11,-9.18 9.18,-9.18h85.43c9.22,0 16.7,-7.47 16.7,-16.7v-66.78C345.04,313.56 337.57,306.09 328.35,306.09z"
+      android:fillColor="#5CB8E6"/>
+  <path
+      android:pathData="M233.74,306.09H83.48c-5.72,0 -11.3,0.58 -16.7,1.68v15.02c0,46.03 37.45,83.48 83.48,83.48s83.48,-37.45 83.48,-83.48V306.09z"
+      android:fillColor="#E6F3FF"/>
+  <path
+      android:pathData="M150.26,306.09v100.17c46.03,0 83.48,-37.45 83.48,-83.48v-16.7H150.26z"
+      android:fillColor="#B8C9D9"/>
+  <path
+      android:pathData="M495.3,506.43H328.35c-9.22,0 -16.7,-7.47 -16.7,-16.7V172.52c0,-55.24 44.94,-100.17 100.17,-100.17S512,117.29 512,172.52v317.22C512,498.96 504.52,506.43 495.3,506.43z"
+      android:fillColor="#FFAA00"/>
+  <path
+      android:pathData="M495.3,506.43c9.22,0 16.7,-7.47 16.7,-16.7V172.52c0,-55.24 -44.94,-100.17 -100.17,-100.17v434.09H495.3z"
+      android:fillColor="#F28D00"/>
+  <path
+      android:pathData="M150.26,372.87c-27.62,0 -50.09,-22.47 -50.09,-50.09v-39.09c0,-9.22 7.47,-16.7 16.7,-16.7h66.78c9.22,0 16.7,7.47 16.7,16.7v39.09C200.35,350.4 177.88,372.87 150.26,372.87z"
+      android:fillColor="#FFCC99"/>
+  <path
+      android:pathData="M183.65,267h-33.39V372.87c27.62,0 50.09,-22.47 50.09,-50.09v-39.09C200.35,274.47 192.87,267 183.65,267z"
+      android:fillColor="#F2AC79"/>
+  <path
+      android:pathData="M428.52,306.09H395.13c-9.22,0 -16.7,-7.47 -16.7,-16.7c0,-9.22 7.47,-16.7 16.7,-16.7h33.39c9.22,0 16.7,7.47 16.7,16.7C445.22,298.61 437.74,306.09 428.52,306.09z"
+      android:fillColor="#007AB3"/>
+  <path
+      android:pathData="M411.83,306.09h16.7c9.22,0 16.7,-7.47 16.7,-16.7c0,-9.22 -7.47,-16.7 -16.7,-16.7h-16.7V306.09z"
+      android:fillColor="#005E80"/>
+  <path
+      android:pathData="M428.52,239.3H395.13c-9.22,0 -16.7,-7.47 -16.7,-16.7s7.47,-16.7 16.7,-16.7h33.39c9.22,0 16.7,7.47 16.7,16.7S437.74,239.3 428.52,239.3z"
+      android:fillColor="#007AB3"/>
+  <path
+      android:pathData="M411.83,239.3h16.7c9.22,0 16.7,-7.47 16.7,-16.7s-7.47,-16.7 -16.7,-16.7h-16.7V239.3z"
+      android:fillColor="#005E80"/>
+  <path
+      android:pathData="M150.26,306.09c-64.44,0 -116.87,-52.43 -116.87,-116.87c0,-9.22 7.47,-16.7 16.7,-16.7h200.35c9.22,0 16.7,7.47 16.7,16.7C267.13,253.66 214.7,306.09 150.26,306.09z"
+      android:fillColor="#FFEECC"/>
+  <path
+      android:pathData="M250.43,172.52H150.26v133.57c64.44,0 116.87,-52.43 116.87,-116.87C267.13,180 259.65,172.52 250.43,172.52z"
+      android:fillColor="#FFD9B3"/>
+  <path
+      android:pathData="M33.39,189.22c0,27.62 22.47,50.09 50.09,50.09h33.39c12.82,0 24.52,-4.85 33.39,-12.79c8.87,7.95 20.57,12.79 33.39,12.79h33.39c27.62,0 50.09,-22.47 50.09,-50.09v-33.39H33.39V189.22z"
+      android:fillColor="#99DDFF"/>
+  <path
+      android:pathData="M150.26,226.51c8.87,7.95 20.57,12.79 33.39,12.79h33.39c27.62,0 50.09,-22.47 50.09,-50.09v-33.39h-116.87V226.51z"
+      android:fillColor="#73C0E6"/>
+  <path
+      android:pathData="M150.26,105.74c-27.62,0 -50.09,-22.47 -50.09,-50.09s22.47,-50.09 50.09,-50.09s50.09,22.47 50.09,50.09S177.88,105.74 150.26,105.74z"
+      android:fillColor="#E6F3FF"/>
+  <path
+      android:pathData="M150.26,5.57v100.17c27.62,0 50.09,-22.47 50.09,-50.09S177.88,5.57 150.26,5.57z"
+      android:fillColor="#B8C9D9"/>
+  <path
+      android:pathData="M265.96,139.19c-0,-0.02 -0,-0.04 -0.01,-0.06c-3.6,-25.5 -15.48,-48.92 -34.33,-67.19c-22.88,-22.2 -53.08,-33.87 -85,-32.92c-30.45,0.94 -58.92,13.5 -80.18,35.38c-17.38,17.89 -28.39,40.44 -31.85,64.74c-0,0.02 -0.01,0.05 -0.01,0.07L51.2,155.83h198.12L265.96,139.19z"
+      android:fillColor="#007AB3"/>
+  <path
+      android:pathData="M265.96,139.19c-0,-0.02 -0,-0.04 -0.01,-0.06c-3.6,-25.5 -15.48,-48.92 -34.33,-67.19c-22.02,-21.36 -50.79,-32.96 -81.36,-32.97v116.86h99.06L265.96,139.19z"
+      android:fillColor="#005E80"/>
+  <path
+      android:pathData="M267.13,155.82v16.71H33.39v-16.7c0,-5.63 0.4,-11.21 1.19,-16.7h231.37C266.74,144.61 267.13,150.18 267.13,155.82z"
+      android:fillColor="#E6F3FF"/>
+  <path
+      android:pathData="M267.13,155.82v16.71h-116.87V139.13h115.69C266.74,144.61 267.13,150.18 267.13,155.82z"
+      android:fillColor="#B8C9D9"/>
+  <path
+      android:pathData="M50.09,406.26c-9.22,0 -16.7,7.47 -16.7,16.7v83.48h33.39v-83.48C66.78,413.74 59.31,406.26 50.09,406.26z"
+      android:fillColor="#5CB8E6"/>
+</vector>
