+++ Snowboard/app/src/main/res/layout/fragment_splash_screen.xml
@@ -2,17 +2,19 @@
 <layout>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
+        android:background="@color/crimson_red"
         tools:context=".fragments.SplashScreenFragment">
 
         <ImageView
+            android:id="@+id/imageLogo"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:scaleType="center"
-            android:src="@drawable/splash_screen" />
+            android:src="@drawable/ic_snowboard_logo" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
